+++ src/emsuite/__init__.py
@@ -2,22 +2,16 @@
 
 __version__ = "1.3.0"
 
-from emsuite import api
-from emsuite.api import tune
-from emsuite.config import load_config, parse_config_file
 from emsuite.coupled import run_coupled_calculation
-from emsuite.geometry import Geometry
 from emsuite.inputs import CoupledInput, PotentialInput, SurfaceInput, TuningInput
 from emsuite.potential import run_potential_calculation
 from emsuite.results import CoupledResult, PotentialResult, SurfaceResult, TuningResult
-from emsuite.surface import generate_surface, run_surface_calculation
-from emsuite.tuning import get_tuning_parameters
-from emsuite.tuning import main as run_tuning
+from emsuite.surface import run_surface_calculation
+from emsuite.tuning import run_tuning_calculation
 
 __all__ = [
     "CoupledInput",
     "CoupledResult",
-    "Geometry",
     "PotentialInput",
     "PotentialResult",
     "SurfaceInput",
@@ -25,14 +19,8 @@ __all__ = [
     "TuningInput",
     "TuningResult",
     "__version__",
-    "api",
-    "generate_surface",
-    "get_tuning_parameters",
-    "load_config",
-    "parse_config_file",
     "run_coupled_calculation",
     "run_potential_calculation",
     "run_surface_calculation",
-    "run_tuning",
-    "tune",
+    "run_tuning_calculation",
 ]